+++ opencv/src/gen/java/org/bytedeco/opencv/opencv_mcc/CCheckerDraw.java
@@ -2,6 +2,7 @@
 
 package org.bytedeco.opencv.opencv_mcc;
 
+import org.bytedeco.javacpp.annotation.Index;
 import java.nio.*;
 import org.bytedeco.javacpp.*;
 import org.bytedeco.javacpp.annotation.*;